feat: Add React-Router for better page routing and navigation. Organize project for better readibility and scalability.
