+++ package.json
@@ -14,6 +14,7 @@
     "lucide-react": "^0.511.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
+    "react-router": "^7.6.1",
     "tailwindcss": "^4.1.7"
   },
   "devDependencies": {
