+++ src/Routes.tsx
@@ -1,0 +1,35 @@
+import { createBrowserRouter } from "react-router";
+import App from "./App";
+import { Products } from "./pages/Products";
+import { Home } from "./pages/Home";
+import { Process } from "./pages/Process";
+import { About } from "./pages/About";
+import { Contact } from "./pages/Contact";
+
+export const router = createBrowserRouter([
+  {
+    element: <App />,
+    children: [
+      {
+        path: "/",
+        element: <Home />,
+      },
+      {
+        path: "/products",
+        element: <Products />,
+      },
+      {
+        path: "/process",
+        element: <Process />,
+      },
+      {
+        path: "/about",
+        element: <About />,
+      },
+      {
+        path: "/contact",
+        element: <Contact />,
+      },
+    ],
+  },
+]);
