+++ src/pages/Home.tsx
@@ -1,0 +1,5 @@
+import { HeroSection } from "../components/Hero";
+
+export function Home() {
+  return <HeroSection />;
+}
